[refactor/#73]:스크롤throttle 적용 및 hook 분리
